test!: I FIXED THE PROBLEM
It appiers to be the case that PostgreSQL (or SQL engines in common)
are not concurent-safe. And because of that if two SQL queries `CREATE
EXTENTION IF NOT EXISTS "uuid-ossp";` was run simultaneously than
"uuid-ossp" extention was not imported correctly and I got error message:
`pq: function uuid_generate_v4() does not exist`.

The solution was to simply run `-p 1` (means 'set parallel runs to 1') option in `go test ./...` command,
like so: `go test -p 1 ./...`. It makes Go to run test sequentially and
prevent error.